Force a higher runtime dependency on libidn2
We use enums from a higher version of libidn2 that aren't
automatically detected by shlibdeps.  This means we need to explicitly
declare a higher version in the runtime dependencies than shlibdeps
would otherwise do.  At a minimum, we need libidn2 (>= 0.14), but as
Tim says in the libpsl 0.17 release notes, libidn2 0.16 fixes two
bugs, so we'll just skip straight to that.